# core-orchestrator-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-orchestrator API
  description: Core processes orchestrator
  contact:
    name: Catalis Development
    url: 'https://getfirefly.io'
    email: dev@catalis.com
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0'
  version: 1.0.0
servers:
  - url: 'http://localhost:8081'
    description: Local Development Server
...
paths:
  /api/v1/notifications/sms/validate-code:
    post:
      tags:
        - SMS Notifications
      summary: Validate SMS verification code
      description: Starts a process to validate a verification code sent via SMS
      operationId: validateSMSVerificationCode
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateCodeRequest'
        required: true
      responses:
        '200':
          description: Validation process started successfully
          content:
            application/json: {}
        '500':
          description: Internal server error
          content:
            application/json: {}
  /api/v1/notifications/sms/send-verification:
    post:
      tags:
        - SMS Notifications
      summary: Send verification SMS
      description: Starts a process to send a verification SMS to the specified recipient
      operationId: sendVerificationSMS
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
        required: true
      responses:
        '200':
          description: Process started successfully
          content:
            application/json: {}
        '500':
          description: Internal server error
          content:
            application/json: {}
  /api/v1/notifications/email/validate-code:
    post:
      tags:
        - Email Notifications
      summary: Validate email verification code
      description: Starts a process to validate a verification code sent via email
      operationId: validateEmailVerificationCode
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateCodeRequest'
        required: true
      responses:
        '200':
          description: Validation process started successfully
          content:
            application/json: {}
        '500':
          description: Internal server error
          content:
            application/json: {}
  /api/v1/notifications/email/send-verification:
    post:
      tags:
        - Email Notifications
      summary: Send verification email
      description: Starts a process to send a verification email to the specified recipient
      operationId: sendVerificationEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
        required: true
      responses:
        '200':
          description: Process started successfully
          content:
            application/json: {}
        '500':
          description: Internal server error
          content:
            application/json: {}
...
components:
  schemas:
    ValidateCodeRequest:
      type: object
      properties:
        idOperation:
          type: integer
          format: int64
        code:
          type: string
      description: Verification code validation request details
    NotificationRequest:
      type: object
      properties:
        to:
          type: string
        idOperation:
          type: integer
          format: int64
      description: Email notification request details
...